requirements.txt
nest-asyncio
requests[socks]
PyYAML
InquirerPy
rich
